examples/gemini/requirements.txt
google-genai>=1.30.0
google-adk>=1.18.0
langfuse>=3.14.0
python-dotenv>=1.0.1
